requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
confection==0.0.3
cymem==2.0.7
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
jsonschema==4.17.1
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.9
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pathy==0.10.0
Pillow==9.3.0
preshed==3.0.8
protobuf==3.20.3
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.4.0/pt_core_news_lg-3.4.0-py3-none-any.whl
pyarrow==10.0.1
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
streamlit==1.15.1
streamlit-ace==0.1.1
thinc==8.1.5
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wasabi==0.10.1
watchdog==2.1.9
zipp==3.11.0
